Add files via upload
This .py script is like the elasticdump tool which is built as a pure python script. 
Since the elasticdump is built and installed via node, it narrowed the available scenes.
Also it adds an option to specify the fields you queried, so to export a much smaller data used to update.
And it holds a considerable speed in both data downloading and uploading.